Solve day 11 part 2
Full credit to reddit for the "inspiraition". The missing piece was
remembering that we only care about the count. This sped up the entire
function since it didn't have to generate so many values and iterate
over them. Other than that, I think I had the right idea barring
overcomplicating the memoization in my previous implementations.